# Speed up lookupAnalysis and implement shadowing detection