--- a/internal/zinc-core/src/main/scala/sbt/internal/inc/Compile.scala
+++ b/internal/zinc-core/src/main/scala/sbt/internal/inc/Compile.scala
@@ -70,7 +70,7 @@ object IncrementalCompile {
     }
   def getExternalAPI(lookup: Lookup): (File, String) => Option[AnalyzedClass] =
     (file: File, binaryClassName: String) =>
-      lookup.lookupAnalysis(file, binaryClassName) flatMap {
+      lookup.lookupAnalysis(binaryClassName) flatMap {
         case (analysis: Analysis) =>
           val sourceClassName = analysis.relations.binaryClassName.reverse(binaryClassName).headOption
           sourceClassName flatMap analysis.apis.internal.get