--- a/internal/zinc-core/src/main/scala/sbt/internal/inc/IncrementalCommon.scala
+++ b/internal/zinc-core/src/main/scala/sbt/internal/inc/IncrementalCommon.scala
@@ -370,8 +370,8 @@ private[inc] abstract class IncrementalCommon(log: sbt.util.Logger, options: Inc
             }
           }
 
-        lookup.lookupAnalysis(dependsOn).isEmpty &&
-          (if (skipClasspathLookup) fileModified(dependsOn, dependsOn) else dependencyModified(dependsOn))
+        // lookup.lookupAnalysis(dependsOn).isEmpty &&
+        (if (skipClasspathLookup) fileModified(dependsOn, dependsOn) else dependencyModified(dependsOn))
 
       }