--- a/internal/zinc-core/src/main/scala/sbt/internal/inc/Lookup.scala
+++ b/internal/zinc-core/src/main/scala/sbt/internal/inc/Lookup.scala
@@ -9,6 +9,11 @@ import xsbti.compile.{ CompileAnalysis, ExternalHooks }
  * an external (for another subproject) Analysis instance.
  */
 trait Lookup extends ExternalLookup {
+  /**
+   * Returns the current classpath if the classpath has changed from the last compilation.
+   */
+  def changedClasspath: Option[Vector[File]]
+
   def analyses: Vector[CompileAnalysis]
 
   /**