--- a/internal/zinc-persist/src/main/scala/sbt/internal/inc/TextAnalysisFormat.scala
+++ b/internal/zinc-persist/src/main/scala/sbt/internal/inc/TextAnalysisFormat.scala
@@ -354,6 +354,7 @@ object TextAnalysisFormat {
     object Headers {
       val outputMode = "output mode"
       val outputDir = "output directories"
+      val classpathOptions = "classpath options"
       val compileOptions = "compile options"
       val javacOptions = "javac options"
       val compilerVersion = "compiler version"
@@ -377,6 +378,7 @@ object TextAnalysisFormat {
 
       writeSeq(out)(Headers.outputMode, mode :: Nil, identity[String])
       writeMap(out)(Headers.outputDir, outputAsMap, fileToString, fileToString)
+      writeSeq(out)(Headers.classpathOptions, setup.options.classpath, fileToString)
       writeSeq(out)(Headers.compileOptions, setup.options.scalacOptions, identity[String])
       writeSeq(out)(Headers.javacOptions, setup.options.javacOptions, identity[String])
       writeSeq(out)(Headers.compilerVersion, setup.compilerVersion :: Nil, identity[String])
@@ -390,6 +392,7 @@ object TextAnalysisFormat {
       def s2b(s: String): Boolean = s.toBoolean
       val outputDirMode = readSeq(in)(Headers.outputMode, identity[String]).headOption
       val outputAsMap = readMap(in)(Headers.outputDir, stringToFile, stringToFile)
+      val classpathOptions = readSeq(in)(Headers.classpathOptions, stringToFile)
       val compileOptions = readSeq(in)(Headers.compileOptions, identity[String])
       val javacOptions = readSeq(in)(Headers.javacOptions, identity[String])
       val compilerVersion = readSeq(in)(Headers.compilerVersion, identity[String]).head
@@ -418,7 +421,7 @@ object TextAnalysisFormat {
         case None => throw new ReadException("No output mode specified")
       }
 
-      new MiniSetup(output, new MiniOptions(compileOptions.toArray, javacOptions.toArray), compilerVersion,
+      new MiniSetup(output, new MiniOptions(classpathOptions.toArray, compileOptions.toArray, javacOptions.toArray), compilerVersion,
         xsbti.compile.CompileOrder.valueOf(compileOrder), nameHashing, skipApiStoring, extra.toArray)
     }
   }