… internal/compiler-interface/src/main/datatype/incremental.json
… internal/compiler-interface/src/main/datatype/other.json
… internal/zinc-core/src/main/scala/sbt/internal/inc/Stamp.scala
… internal/zinc-core/src/test/scala/sbt/TestAnalyzingCompiler.scala
… internal/zinc-core/src/test/scala/sbt/inc/TestCaseGenerators.scala
… internal/zinc-persist/src/test/scala/sbt/inc/TextAnalysisFormatSpecification.scala
… zinc/src/main/scala/sbt/internal/inc/MixedAnalyzingCompiler.scala
… zinc/src/test/scala/sbt/inc/MultiProjectIncrementalSpec.scala