This PR fixes #166 and sbt/sbt#2525.

## lookupAnalysis speedup

d8c91eb0dae51d8b9f507b0e1d1384898108e33c is a port of https://github.com/gkossakowski/sbt/commit/c45f580b781aaba16e823a8db6f9599f2ca935ae, which should address the performance improvement mentioned in sbt/sbt#2525.

## Relations rename

c71f1f6933c1f5bd2ab852b39dcc83e80b95da6d and 488c977ce3eff9f21e113d41922294ed6802a343 are minor refactoring. The term "binary" is used in multiple ways, so to disambiguate `binaryDeps` are now called `libraryDeps`, and `binaryClassName` is now called `productClassName`.

## full tracking of class names coming from libraries

747015f814a2a0fb65dedc9426dc53add860741f adds a new relation called `libraryClassName`, which tracks JAR -> fully qualified class names that comes up in dependencies. Stamp does track class name, but it only keeps the last one.

## shadow detection fix

ae5960180cbfee4ee78d426cc7f4679aff64b9f0 improves the existing shadow detection facility (`externalBinaryModified`) but using the newly added `libraryClassName` relation. If we care only about the correctness that should be ok.

## perf improvement of the shadow detection

4417092fca56b51cde2559861f85a9cb3878e0ac 
improves the shadow detection perf by using the new `lookupAnalysis`, which scans `Analysis` objects in the order of the classpath instead of doing the io-based `lookup.lookupOnClasspath`.

/review @gkossakowski, @romanowski, @dwijnand