@eed3si9n 2016-09-28
I talked to @gkossakowski during the day (Eastern) and we confirmed a few points.
- Stamps only tracks one class name so it's inadequate. I've demonstrated its shortcoming in #179.
- Shadow detection is already sort of implemented as `binaryDepChanges`, which calls `externalBinaryModified`, and that I should extend its behavior.
- Perf improvement can be gained by replacing the classpath lookup in `externalBinaryModified` with Analysis scanning.

The last 3 commits implement those ideas.