@gkossakowski 2016-09-29
Did you try to measure performance gains using the sample project from sbt/sbt#2525?