@eed3si9n 2016-09-29
Yes. The speedup I'm seeing around 1.5x to 2x, but that's just for source changes.
If we test for binary change scanning we should see more improvements.

### Windows with hdd

Before:

```
> check
Running compile took 3233ms
> check
Running compile took 2750ms
> check
Running compile took 2650ms
> check
Running compile took 2647ms
> check
Running compile took 2678ms
```

Median: 2678ms

After:

```
> check
Running compile took 1792ms
> check
Running compile took 1547ms
> check
Running compile took 1451ms
> check
Running compile took 1461ms
> check
Running compile took 1471ms
```

Median: 1471ms

### Mac with SSD

Before:

```
> check
running compile took 570.49902 ms
> check
running compile took 927.169266 ms
> check
running compile took 295.587101 ms
> check
running compile took 305.026124 ms
> check
running compile took 306.711616 ms
```

Median: 306ms


After:

```
> check
running compile took 988.027559 ms
> check
running compile took 289.089768 ms
> check
running compile took 190.252249 ms
> check
running compile took 214.043261 ms
> check
running compile took 197.949601 ms
```

Median: 214ms