@eed3si9n 2016-10-05
btw I'm out in Texas this week for Reactive Summit + Lightbend tech meetings, so my response is slower than usual. I'll tackle the suggested changes when I get back.