@gkossakowski 2016-10-05
Sure. Thanks for letting me know.

On 5 October 2016 at 05:22, eugene yokota <notifications@github.com> wrote:

> btw I'm out in Texas this week for Reactive Summit + Lightbend tech
> meetings, so my response is slower than usual. I'll tackle the suggested
> changes when I get back.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/sbt/zinc/pull/180#issuecomment-251572611>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAAQeUXSZDe6s2GdKxUOZbXPZPuDpcZBks5qwxf8gaJpZM4KHdWb>
> .
>



-- 
gkk