@gkossakowski 2016-10-05
Overall, the changes look good. I'm not sure about the renaming of `binary` to `library`. I don't see the ambiguity you're trying to resolve. Can you give an example?

That is to say, I'm asking more for documentation purposes than debating the merit of the change itself.