@eed3si9n 2016-10-19
@gkossakowski @romanowski I've addressed the review comments.